Changed the label of the Record to be of type double, which is what is compatible for a neural network; it was an int prevously.  Now the label has to be continous because the output of my neural network is continuous.

+++ src/Record.java
@@ -5,11 +5,11 @@ public class Record {
 	public double[] getAttrList() {
 		return attrList;
 	}
-	private int label;
-	public int getLabel() {
+	private double label;
+	public double getLabel() {
 			return label;
 	}
-	public Record(double[] attrList, int label) {
+	public Record(double[] attrList, double label) {
 		this.attrList = attrList;
 		this.label = label;
 	}
